+++ app/src/main/java/awesomegic/bank/operation/WithdrawOperation.java
@@ -1,5 +1,7 @@
 package awesomegic.bank.operation;
 
+import java.math.BigDecimal;
+
 import awesomegic.bank.cli.Cli;
 import awesomegic.bank.cli.exceptions.InputException;
 import awesomegic.bank.model.account.BankAccount;
@@ -37,7 +39,7 @@ public class WithdrawOperation implements Operation {
      */
     @Override
     public OperationResult execute() throws InputException, OperationException {
-        double amount = this.cli.readTransactionAmount(TRANSACTION_TYPE);
+        BigDecimal amount = this.cli.readTransactionAmount(TRANSACTION_TYPE);
 
         if (!this.account.isBalanceSufficient(amount)) {
             throw new OperationException(MESSAGE_INSUFFICIENT_BALANCE);
